requirements.txt
Flask==2.3.3
groq==0.9.0
python-dotenv==1.0.0